up8

+++ Server/src/org/crandor/game/node/item/WeightedChanceItem.java
@@ -4,16 +4,16 @@ import org.crandor.tools.RandomFunction;
 
 public class WeightedChanceItem {
     int id,minimum_amount,maximum_amount;
-    public double weight;
+    public int weight;
 
-    public WeightedChanceItem(int id, int minimum_amount, int maximum_amount, double weight){
+    public WeightedChanceItem(int id, int minimum_amount, int maximum_amount, int weight){
         this.id = id;
         this.minimum_amount = minimum_amount;
         this.maximum_amount = maximum_amount;
         this.weight = weight;
     }
 
-    public WeightedChanceItem(int id, int amount, double weight){
+    public WeightedChanceItem(int id, int amount, int weight){
         this(id,amount,amount,weight);
     }
